+++ src/core/editor/pdf_editor.js
@@ -25,7 +25,6 @@ import {
   getInheritableProperty,
   getModificationDate,
   getNewAnnotationsMap,
-  stringToAsciiOrUTF16BE,
 } from "../core_utils.js";
 import { Dict, isName, Name, Ref, RefSet, RefSetCache } from "../primitives.js";
 import { incrementalUpdate, writeValue } from "../writer.js";
@@ -34,6 +33,7 @@ import { stringToBytes, stringToPDFString } from "../../shared/util.js";
 import { AnnotationFactory } from "../annotation.js";
 import { BaseStream } from "../base_stream.js";
 import { StringStream } from "../stream.js";
+import { stringToAsciiOrUTF16BE } from "../string_utils.js";
 
 const MAX_LEAVES_PER_PAGES_NODE = 16;
 const MAX_IN_NAME_TREE_NODE = 64;
