+++ src/core/struct_tree.js
@@ -21,9 +21,10 @@ import {
   warn,
 } from "../shared/util.js";
 import { Dict, isName, Name, Ref, RefSetCache } from "./primitives.js";
-import { lookupNormalRect, stringToAsciiOrUTF16BE } from "./core_utils.js";
 import { BaseStream } from "./base_stream.js";
+import { lookupNormalRect } from "./core_utils.js";
 import { NumberTree } from "./name_number_tree.js";
+import { stringToAsciiOrUTF16BE } from "./string_utils.js";
 
 const MAX_DEPTH = 40;
