+++ test/unit/clitests.json
@@ -49,6 +49,7 @@
     "postscript_spec.js",
     "primitives_spec.js",
     "stream_spec.js",
+    "string_utils_spec.js",
     "struct_tree_spec.js",
     "svg_factory_spec.js",
     "text_layer_spec.js",
